Create display() and work out logic for button presses
Updated the operator function with correct # of arguments to be passed to it.

Initiated logic for display() that will update the screen when buttons are presssed
have implemented logic that will seperate numbered button presses (which are to be displayed)
vs. operator buttons which will have conditions applied to them